Extracted some of the functions
